api-filmes/requirements.txt
anyio==3.6.2
astroid==2.14.2
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
bson==0.5.10
certifi==2022.12.7
click==8.1.3
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
orjson==3.8.6
packaging==23.0
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.16.2
pymongo==4.3.3
pytest==7.2.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.25.0
tomlkit==0.11.6
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
